Homework 05 - fix

+++ src/ru/shendo/homework05/medium/exercise02/IProduct.java
@@ -1,10 +1,12 @@
 package ru.shendo.homework05.medium.exercise02;
 
+import java.math.BigDecimal;
+
 public interface IProduct {
 
     String getName();
 
-    double getPrice();
+    BigDecimal getPrice();
 
     int getQuantity();
 }
